Issue #25: For OS X files, made the configure and Makefile options and flags as similar as possible.
I was actually able to get the extra configure flags down to 0 for 64-bit OS X, since that's our primary development platform.

I was able to eliminate a lot of the flags from the Makefiles which were necessary before.  I think the difference was that I recently upgraded to Xcode 4.4 (from 4.2.x, I think) in which I believe Apple switched from using GCC to LLVM under the hood.  Not entirely certain, but whatever the case it makes for cleaner Makefiles.